ขอมูลแสดงได้ครบแล้ว

+++ src/main/java/com/example/demo/module/web/president/repository/AdministrationRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.module.web.president.entity.AdministrationEntity;
 
 @Repository
-public interface AdministrationRepository extends  JpaRepository<AdministrationEntity, Float>{
- 
+public interface AdministrationRepository extends JpaRepository<AdministrationEntity, Long> {
 }
